backend/requirements.txt
torch
torchvision
datasets
transformers
accelerate
bitsandbytes
sentencepiece
protobuf
pymupdf
langchain
langchain-community
pypdf
pymupdf4llm
numpy
sentence-transformers
chromadb
ollama
openai
scikit-learn
fastapi
fastapi[standard]
typing
pydantic
python-dotenv
python-multipart
huggingface_hub
langchain_experimental
